# Improve GEMM performance by nearly 2x