- Moves tinyblas from `llama.cpp` into `llamafile`.
- Uses by-row 8x parallelization on all GEMM routines.
- Fixes a build regression when using `-DGGML_USE_TINYBLAS`