@mrdomino 2023-12-14
Tested against `llava-v1.5-7b-Q4_K.gguf` at temp 0, seed 1234, prompt:

```
$'\nvoid *memcpy(void *dst, const void *src, size_t size) {\n'
```

Before this change:
```
llama_print_timings:        load time =    1258.33 ms
llama_print_timings:      sample time =      37.44 ms /    53 runs   (    0.71 ms per token,  1415.71 tokens per second)
llama_print_timings: prompt eval time =     478.37 ms /    24 tokens (   19.93 ms per token,    50.17 tokens per second)
llama_print_timings:        eval time =    3260.49 ms /    52 runs   (   62.70 ms per token,    15.95 tokens per second)
llama_print_timings:       total time =    3785.20 ms
```

after:
```
llama_print_timings:        load time =    1250.69 ms
llama_print_timings:      sample time =      37.46 ms /    53 runs   (    0.71 ms per token,  1414.96 tokens per second)
llama_print_timings: prompt eval time =     227.51 ms /    24 tokens (    9.48 ms per token,   105.49 tokens per second)
llama_print_timings:        eval time =    1898.23 ms /    52 runs   (   36.50 ms per token,    27.39 tokens per second)
llama_print_timings:       total time =    2178.08 ms
```