@mrdomino 2023-12-14
Modulo strange divergences as described in #4458 (with the commit in question, tinyblas does not agree with cublas, nor does cublas agree with cpu; however, tinyblas in this commit does (!) agree with tinyblas at HEAD), this is ready to go on my end.