@mrdomino 2023-12-14
(FYI credit goes to @ahgamut for the approach.)